+++ include/omw/utility.h
@@ -18,38 +18,38 @@ namespace omw
     * @{
     */
 
-    //int8_t shiftLeft(int8_t value, size_t n);
-    //int16_t shiftLeft(int16_t value, size_t n);
-    //int32_t shiftLeft(int32_t value, size_t n);
-    //int64_t shiftLeft(int64_t value, size_t n);
-    //uint8_t shiftLeft(uint8_t value, size_t n);
-    //uint16_t shiftLeft(uint16_t value, size_t n);
-    //uint32_t shiftLeft(uint32_t value, size_t n);
-    //uint64_t shiftLeft(uint64_t value, size_t n);
-    //int8_t shiftRight(int8_t value, size_t n);
-    //int16_t shiftRight(int16_t value, size_t n);
-    //int32_t shiftRight(int32_t value, size_t n);
-    //int64_t shiftRight(int64_t value, size_t n);
-    //uint8_t shiftRight(uint8_t value, size_t n);
-    //uint16_t shiftRight(uint16_t value, size_t n);
-    //uint32_t shiftRight(uint32_t value, size_t n);
-    //uint64_t shiftRight(uint64_t value, size_t n);
-    void shiftLeftAssign(int8_t& value, size_t n);
-    void shiftLeftAssign(int16_t& value, size_t n);
-    void shiftLeftAssign(int32_t& value, size_t n);
-    void shiftLeftAssign(int64_t& value, size_t n);
-    void shiftLeftAssign(uint8_t& value, size_t n);
-    void shiftLeftAssign(uint16_t& value, size_t n);
-    void shiftLeftAssign(uint32_t& value, size_t n);
-    void shiftLeftAssign(uint64_t& value, size_t n);
-    void shiftRightAssign(int8_t& value, size_t n);
-    void shiftRightAssign(int16_t& value, size_t n);
-    void shiftRightAssign(int32_t& value, size_t n);
-    void shiftRightAssign(int64_t& value, size_t n);
-    void shiftRightAssign(uint8_t& value, size_t n);
-    void shiftRightAssign(uint16_t& value, size_t n);
-    void shiftRightAssign(uint32_t& value, size_t n);
-    void shiftRightAssign(uint64_t& value, size_t n);
+    void shiftLeftAssign(int8_t& value, unsigned int n);
+    void shiftLeftAssign(int16_t& value, unsigned int n);
+    void shiftLeftAssign(int32_t& value, unsigned int n); /*!< See \ref grp_utility_langSupport_section_bitShiftOp in \ref grp_utility_langSupport. */
+    void shiftLeftAssign(int64_t& value, unsigned int n);
+    void shiftLeftAssign(uint8_t& value, unsigned int n);
+    void shiftLeftAssign(uint16_t& value, unsigned int n);
+    void shiftLeftAssign(uint32_t& value, unsigned int n);
+    void shiftLeftAssign(uint64_t& value, unsigned int n);
+    void shiftRightAssign(int8_t& value, unsigned int n);
+    void shiftRightAssign(int16_t& value, unsigned int n);
+    void shiftRightAssign(int32_t& value, unsigned int n); /*!< See \ref grp_utility_langSupport_section_bitShiftOp in \ref grp_utility_langSupport. */
+    void shiftRightAssign(int64_t& value, unsigned int n);
+    void shiftRightAssign(uint8_t& value, unsigned int n);
+    void shiftRightAssign(uint16_t& value, unsigned int n);
+    void shiftRightAssign(uint32_t& value, unsigned int n);
+    void shiftRightAssign(uint64_t& value, unsigned int n);
+    int8_t shiftLeft(int8_t value, unsigned int n);
+    int16_t shiftLeft(int16_t value, unsigned int n);
+    int32_t shiftLeft(int32_t value, unsigned int n); /*!< See \ref grp_utility_langSupport_section_bitShiftOp in \ref grp_utility_langSupport. */
+    int64_t shiftLeft(int64_t value, unsigned int n);
+    uint8_t shiftLeft(uint8_t value, unsigned int n);
+    uint16_t shiftLeft(uint16_t value, unsigned int n);
+    uint32_t shiftLeft(uint32_t value, unsigned int n);
+    uint64_t shiftLeft(uint64_t value, unsigned int n);
+    int8_t shiftRight(int8_t value, unsigned int n);
+    int16_t shiftRight(int16_t value, unsigned int n);
+    int32_t shiftRight(int32_t value, unsigned int n); /*!< See \ref grp_utility_langSupport_section_bitShiftOp in \ref grp_utility_langSupport. */
+    int64_t shiftRight(int64_t value, unsigned int n);
+    uint8_t shiftRight(uint8_t value, unsigned int n);
+    uint16_t shiftRight(uint16_t value, unsigned int n);
+    uint32_t shiftRight(uint32_t value, unsigned int n);
+    uint64_t shiftRight(uint64_t value, unsigned int n);
 
     // grp_utility_langSupport
     /*! @} */